requirements.txt
arrow==0.14.7
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
aws-requests-auth==0.4.2
aws-sam-translator==1.14.0
aws-xray-sdk==2.4.2
boto==2.49.0
boto3==1.9.220
botocore==1.12.220
certifi==2019.6.16
cffi==1.12.3
cfn-lint==0.23.5
chardet==3.0.4
Click==7.0
cloudwatch-fluent-metrics==0.5.3
cryptography==2.7
DateTime==4.3
docker==4.0.2
docutils==0.15.2
ecdsa==0.13.2
entrypoints==0.3
flake8==3.7.8
future==0.17.1
idna==2.8
importlib-metadata==0.19
Jinja2==2.10.1
jmespath==0.9.4
jsondiff==1.1.2
jsonpatch==1.24
jsonpickle==1.2
jsonpointer==2.0
jsonschema==3.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.2.0
moto==1.3.13
packaging==19.1
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.6
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.2
python-dateutil==2.8.0
python-jose==3.0.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
responses==0.10.6
rsa==4.0
s3transfer==0.2.1
six==1.12.0
sshpubkeys==3.1.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.6.0
zope.interface==4.6.0
